// asr.webapp.v2/src/app/modules/financial/cheque/cheque.types.ts
import { Paging } from "app/core/type/paging/paging.type";
import { BaseModel } from "app/core/type/base/BaseModel";

export class Cheque extends BaseModel {
    chequeId: number;
    chequeNumber: string;
    bankId: number;
    bankName: string;
    bankUrl: string;
    amount: number;
    currencyId: number;
    currencyName: string;
    currencyAbbreviation: string;
    chequeDate: Date;
    chequeLocalDate: string;
    chequeLocalDateClean: string;
    payerId: number;
    payerName: string;
    payerTypeId: number;
    payerTypeName: string;
    recipientId: number;
    recipientName: string;
    recipientTypeId: number;
    recipientTypeName: string;
    beneficiaryId: number;
    beneficiaryName: string;
    beneficiaryTypeId: number;
    beneficiaryTypeName: string;
    firstPayerId: number;
    firstPayerName: string;
    firstPayerTypeId: number;
    firstPayerTypeName: string;
    financialResourceId: number;
    financialResourceName: string;
    chequeTypeId: number;
    chequeTypeName: string;
    fee: number;
    status: number;
    statusDescription: string;
    statusColor: string;
    chequeLocalDateFrom: string;
    chequeLocalDateTo: string;
    amountFrom: number;
    amountTo: number;
    statusIdList: Array<number>;
    payerTypeIdList: Array<number>;
    payerIdList: Array<number>;
    recipientTypeIdList: Array<number>;
    firstPayerTypeIdList: Array<number>;
    recipientIdList: Array<number>;
    bankIdList: Array<number>;
    financialResourceIdList: Array<number>;
    chequeTypeIdList: Array<number>;
    page: Paging;

}
export class ChequeTransaction extends BaseModel {
    chequeTransactionId: number;
    chequeId: number;
    chequeNumber: string;
    bankId: number;
    bankName: string;
    bankUrl: string;
    amount: number;
    currencyId: number;
    currencyName: string;
    currencyAbbreviation: string;
    chequeDate: Date;
    chequeLocalDate: string;
    chequeLocalDateClean: string;
    payerId: number;
    payerName: string;
    payerTypeId: number;
    payerTypeName: string;
    payerFinancialResourceId: number;
    payerFinancialResourceName: string;
    recipientId: number;
    recipientName: string;
    recipientTypeId: number;
    recipientTypeName: string;
    recipientFinancialResourceId: number;
    recipientFinancialResourceName: string;
    status: number;
    statusDescription: string;
    statusColor: string;
}